% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compound_figure.R
\name{compound_figure}
\alias{compound_figure}
\title{Compound figure for decision support}
\usage{
compound_figure(
  model,
  input_table,
  decision_var_name,
  cashflow_var_name,
  model_runs = 10000,
  distribution_method = "smooth_simple_overlay"
)
}
\arguments{
\item{model}{is a user defined model function see the \code{\link[decisionSupport:mcSimulation]{mcSimulation}} function}

\item{input_table}{is a data frame with at least two columns named 'variable' and 'label'. The 'variable column should have one entry for the name of each variable contained in any of the plots. In preparing the figure, the function will replace the variable names with the labels. If the label is missing then the plot will show 'NA' in the place of the variable name. Default is NULL and uses the original variable names.}

\item{decision_var_name}{is the name of the decision outcome named in the \code{return()} of 'model'. This will be used in all plots and analyses except for the cashflow plot. For now this is just one decision option}

\item{cashflow_var_name}{is the name of the cashflow variable named in the \code{return()} of 'model'. This will be used in the cashflow plot}

\item{model_runs}{is the number of time that the model should run. The default is 10,000}

\item{distribution_method}{is the method used in the distribution plot see the \code{\link[decisionSupport:plot_distribution]{plot_distribution}} function}
}
\description{
Simple compound figure of model results and analyses of a binary decision (do or do not do). The figure includes the distribution of the expected outcome, the expected cashflow, as well as the variable importance and the value of information
}
