% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate.R
\name{corMat<-}
\alias{corMat<-}
\title{Replace correlation matrix.}
\usage{
corMat(x) <- value
}
\arguments{
\item{x}{a distribution.}

\item{value}{\code{numeric matrix}: new correlation matrix.}
}
\description{
Replace the correlation matrix.
}

