% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate.R
\name{corMat.estimate}
\alias{corMat.estimate}
\title{Return the correlation matrix of an \code{estimate} object.}
\usage{
\method{corMat}{estimate}(rho)
}
\arguments{
\item{rho}{an \code{\link{estimate}} object.}
}
\description{
This function returns the full correlation matrix of an \code{\link{estimate}} object.
}
\seealso{
\code{\link{estimate}}, \code{\link{row.names.estimate}}, \code{\link{names.estimate}}
}

