% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\docType{package}
\name{decisionSupport}
\alias{decisionSupport}
\alias{decisionSupport-package}
\title{Quantitative Support of Decision Making under Uncertainty}
\description{
The decisionSupport package supports the Monte Carlo simulation
of different decisions.
}
\details{
The package is subdivided in the Monte Carlo simulation and analysis,
the Value of Information Analysis (VIA) and Partial Least Squares (PLS) analysis
and Variable Importance in Projection (VIP).
}
\references{
Hubbarrd, Douglas W., How to Measure Anything? - Finding the Value of "Intangibles" in Business,
  John Wiley & Sons, Hoboken, New Jersey, 2014, 3rd Ed, \url{http://www.howtomeasureanything.com/}.
}

