% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/empirical_EVPI.R
\name{empirical_EVPI}
\alias{empirical_EVPI}
\title{Expected value of perfect information (EVPI) for many variables. This
is a wrapper for the EVPI_empi function}
\usage{
empirical_EVPI(mc, first_out_var, fileformat = NA, outfolder = NA,
  write_table = FALSE, scale_results = TRUE, legend_table = NULL,
  output_legend_table = NULL)
}
\arguments{
\item{mc}{output table from a Monte Carlo simulation,
e.g. as realized with the decisionSupport package}

\item{first_out_var}{name of the column in the mc table that contains the first output
variable. Information Values are computed for variables in all earlier columns.}

\item{fileformat}{format of the output file. This can only be "png" for output as a file
or NA for plotting within R.}

\item{outfolder}{folder where the outputs should be saved (this is optional).}

\item{write_table}{boolean parameter indicating whether an output table should be written.}

\item{scale_results}{boolean parameter indicating whether numbers in the output should be
expressed in thousands, millions, billions etc.}

\item{legend_table}{data.frame containing translations of input variable names into
labels for the EVPI figure}

\item{output_legend_table}{data.frame containing translations of output variable names into
labels for the EVPI figure}
}
\value{
list of as many elements as there are output variables in the Monte Carlo table:
  each element refers to one of the output variables and contains a data.frame with four
  columns:
  (1) variable - the input variable names
  (2) expected_gain - expected gain when project is
  implemented, without knowing the value of the test
  variable
  (3) EVPI_do - the Expected Value of Perfect Information on the respective input variable,
  if the analysis suggests that the expected value of the decision is likely positive
  (e.g. the project should be done)
  (4) EVPI_dont - the Expected Value of Perfect Information on the respective input variable,
  if the analysis suggests that the expected value of the decision is likely negative
  (e.g. the project should not be done)
}
\description{
This function computes the Value of Information for many variables
contained in the output table of a Monte Carlo simulation. It applies the
\code{\link{EVPI_empi}} function to compute the amount of money that a
rational decision-maker should be willing to pay for perfect information
on each model input variable. See the documentation of the 
\code{\link{EVPI_empi}} function for more details.
}
\examples{


### In the following example, the sign of the calculation is entirely
### determined by the variable indep1, so this should be expected to have
### a high EVPI. Variable indep2 doesn't affect the sign of the output,
### so it should not have information value.

montecarlo <- data.frame(indep1 = rnorm(1000), indep2 = rlnorm(1000))
montecarlo[, 'output1'] <- montecarlo[, 'indep1'] * montecarlo[, 'indep2']
montecarlo[, 'output2'] <- (montecarlo[, 'indep1'] * (montecarlo[, 'indep2']) + 5)

empirical_EVPI(montecarlo,"output1")
}
\author{
Eike Luedeling, Katja Schiffers
}
\keyword{"Value}
\keyword{Information"}
\keyword{of}
