% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate.R
\name{estimate}
\alias{estimate}
\title{Create a multivariate estimate object}
\usage{
estimate(..., correlation_matrix = NULL)
}
\arguments{
\item{...}{arguments that can be coerced to a data frame comprising the base of the estimate.
Mandatory columns are \code{distribution}, \code{lower} and \code{upper}.}

\item{correlation_matrix}{\code{numeric matrix}: containing the correlations of the variables.}
}
\value{
An object of class \code{estimate} which is a list whith components \code{base} and \code{correlation_matrix}:
  \describe{
    \item{\code{$base}}{
    is a \code{\link{data.frame}} with mandatory column \code{distribution}. The \code{\link{row.names}} are the
    names of the variables. Each row has the properties of an \code{\link{estimate1d}}.
    }
    \item{\code{$correlation_matrix}}{
     is a symmetric matrix with row and column names being the subset oft he variables supplied
     in \code{base} which are correlated. Its elements are the corresponding correlations.
     }
  }
}
\description{
This function creates an object of class \code{estimate}. It extends the univariate estimate
\code{\link{estimate1d}} to the multivariate case. This includes the description of correlations
between the different variables.
}
\details{
The parameters in \code{...} provide the base information of an estimate.
\subsection{The structure of the estimate base information (mandatory)}{
   Mandatory columns:
   \tabular{lll}{
     Column name         \tab  R-type    \tab Explanation\cr
     \code{distribution} \tab  \code{character} \tab  Distribution types \cr
     \code{variable}     \tab  \code{character} \tab  Variable names
   }
}
}
\seealso{
\code{\link{row.names.estimate}}, \code{\link{names.estimate}}, \code{\link{corMat}}, \code{\link{estimate_read_csv}},
\code{\link{estimate_write_csv}}, \code{\link{random.estimate}}, \code{\link{estimate1d}}
}

