% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate1d.R
\name{estimate1d}
\alias{as.estimate1d}
\alias{estimate1d}
\title{Create an 1-dimensional estimate object}
\usage{
estimate1d(distribution, lower, upper, ...)

as.estimate1d(x, ...)
}
\arguments{
\item{distribution}{\code{character}; A character string that defines the type of the univariate
parametric distribution.}

\item{lower}{\code{numeric}; lower bound of the 90\% confidence intervall, i.e the 5\%-quantile
of this estimate.}

\item{upper}{\code{numeric}; upper bound of the 90\% confidence intervall, i.e the 95\%-quantile
of this estimate.}

\item{...}{arguments that can be coerced to a data frame (ToDo: or list?) comprising the base of the estimate.}

\item{x}{an object to be transformed to class \code{estimate1d}.}
}
\value{
An object of type \code{estimate1d} which is a list with at least (!) the elements:
   \tabular{lll}{
     \bold{Element}      \tab  \bold{R-type}    \tab \bold{Explanation}\cr
     \code{distribution} \tab  \code{character} \tab  Distribution type of the estimate \cr
     \code{lower}        \tab  \code{numeric}   \tab   5\%-quantile of the estimate\cr
     \code{median}       \tab  \code{numeric}   \tab  50\%-quantile of the estimate\cr
     \code{upper}        \tab  \code{numeric}   \tab  95\%-quantile of the estimate
   }
   \subsection{The \code{median}}{
     Note that the \emph{\code{median}} is a mandatory element of an \code{estimate1d}, although it
     is not necessary as input. In case that no element \code{median} is provided, the default is
     \code{median=mean(c(lower, upper))}. If no median shall be available it has to be set actively
     to \code{NULL}!
   }
}
\description{
This function creates an object of class \code{estimate1d}.

This function tries to transform an object to class \code{estimate1d}.
}
\details{
The parameters in \code{...} provide the base information of an estimate.
\subsection{The structure of the estimate}{ \cr
   Mandatory elements:
   \tabular{lll}{
     \bold{Element}      \tab  \bold{R-type}    \tab \bold{Explanation}\cr
     \code{distribution} \tab  \code{character} \tab  Distribution type of the estimate \cr
     \code{lower}        \tab  \code{numeric}   \tab   5\%-quantile of the estimate\cr
     \code{upper}        \tab  \code{numeric}   \tab  95\%-quantile of the estimate
   }
Optional elements:
    \tabular{lll}{
     \bold{Element}      \tab  \bold{R-type}    \tab \bold{Explanation}\cr
     \code{variable}     \tab  \code{character} \tab  Variable name\cr
     \code{median}       \tab  \code{numeric}   \tab  50\%-quantile of the estimate\cr
     \code{method}       \tab  \code{character} \tab  Method for calculation of distribution parameters
   }
 }
}
\seealso{
\code{\link{random.estimate1d}}
}

