% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate_read_csv_old.R
\name{estimate_read_csv_old}
\alias{estimate_read_csv_old}
\title{Read an Estimate from CSV - File (depreciated standard).}
\usage{
estimate_read_csv_old(fileName, strip.white = TRUE, ...)
}
\arguments{
\item{fileName}{Name of the file containing the marginal information of the estimate that should be read.}

\item{strip.white}{logical. Allows the stripping of leading and trailing white space from unquoted character fields
(numeric fields are always stripped). See \code{\link[base]{scan}} for further details (including the exact meaning of 'white space'),
 remembering that the columns may include the row names.}

\item{...}{Further parameters to be passed to \code{\link[utils]{read.csv}}.}
}
\value{
An object of type \code{\link{estimate}}.
}
\description{
This function reads an \code{\link{estimate}} from the specified csv files. In this context, an estimate of a variable is
defined by its distribution type, its 90\%-confidence interval \code{[lower,upper]} and its correlation to other variables.
#ToDo: Implement characterization of distribution by mean and sd. Eventually, also by other quantiles.
}
\details{
An estimate might consists of uncorrelated and correlated variables. This is reflected in the input file structure, which
is described in the following.
}
\section{CSV input file structures}{

The estimate is read from one or two csv files: the basic csv file which is mandatory and the correlation csv file which is optional.
The basic csv file contains the definition of the distribution of all variables ignoring potential correlations. The correlation csv
file only defines correlations.
\subsection{The structure of the basic input file (mandatory)}{
   File name structure: \code{<basic-filename>.csv}\cr
   Mandatory columns:
   \tabular{lll}{
     Column name         \tab  R-type    \tab Explanation\cr
     \code{lower}        \tab  \code{numeric}   \tab  ToDo \cr
     \code{upper}        \tab  \code{numeric}   \tab  ToDo \cr
     \code{distribution} \tab  \code{character} \tab  ToDo \cr
     \code{variable}     \tab  \code{character} \tab  ToDo
   }
   Optional columns:
   \tabular{lll}{
     Column name         \tab  R-type  \tab  Explanation \cr
     \code{description}  \tab  \code{character}  \tab  ToDo\cr
     \code{median}       \tab  \code{numeric}    \tab  ToDo\cr
     \code{start}        \tab  \code{integer}    \tab  ToDo\cr
     \code{end}          \tab  \code{integer}    \tab  ToDo\cr
     \code{indicator}    \tab  \code{logical}    \tab  ToDo
   }
   Columns without names are ignored. Rows where the \code{variable} field is empty are also dropped.
}
\subsection{The structure of the correlation file (optional)}{
   File name structure: \code{<basic-filename>.csv_correlations.csv}\cr
   #ToDo
}
}
\seealso{
\code{\link{estimate_read_csv}}, \code{\link[utils]{read.csv}}, \code{\link{estimate}}
}

