% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{estimate_write_csv}
\alias{estimate_write_csv}
\title{Write an Estimate to CSV - File.}
\usage{
estimate_write_csv(estimate, fileName, varNamesAsColumn = TRUE,
  quote = FALSE, ...)
}
\arguments{
\item{estimate}{\code{character}. Ouput file name which must end with \code{.csv}.}

\item{varNamesAsColumn}{\code{logical}; If \code{TRUE} the variable names will be written as a
separate column, otherwise as row names.}

\item{...}{Further parameters to be passed to \code{\link[utils]{write.csv}}.}

\item{estimate}{Estimate object to write to file \code{fileName}.}
}
\value{
An object of type \code{\link{estimate}}.
}
\description{
This function writes an \code{\link{estimate}} to the specified csv file(s).
}
\seealso{
\code{\link{estimate_read_csv}}, \code{\link{estimate}}, \code{\link[utils]{write.csv}}
}

