% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/globalNames2data.frameNames.R
\name{globalNames2data.frameNames}
\alias{globalNames2data.frameNames}
\title{Transform model function variable names: global to data.frame names.}
\usage{
globalNames2data.frameNames(modelFunction, globalNames)
}
\arguments{
\item{modelFunction}{a function which body contains global variables. The function must not
contain any arguments.}

\item{globalNames}{a \code{character} vector containing the names of the global variables that
shall be transformed.}
}
\value{
The transformed function which is of the form:
 \preformatted{
   function(x){
     ...
     <expression with x$variable1>
     ...
   }
 }
}
\description{
The variable names of a function are transformed from global variable names to data.frame names
of the form \code{x$<globalName>}.
}
\details{
The input function must be of the form:
 \preformatted{
   modelFunction<-function(){
     ...
     <expression with variable1>
     ...
   }
 }
}
\section{Warning}{

  If there are local functions within the function \code{modelFunction} defined, which arguments
  have identical names to any of the \code{globalNames} the function fails!
}
\examples{
profit1<-function(){
   list(Profit=revenue-costs)
 }
 profit2<-globalNames2data.frameNames(modelFunction=profit1, globalNames=c("revenue", "costs"))
 print(profit2)
 is.function(profit2)
 profit2(data.frame("revenue"=10,"costs"=2))
}
\seealso{
\code{\link{mcSimulation}}, \code{\link{estimate}}
}

