% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{hist.mcSimulation}
\alias{hist.mcSimulation}
\title{Plot histogram of results of a Monte Carlo Simulation.}
\usage{
\method{hist}{mcSimulation}(x, breaks = 100, col = NULL, xlab = NULL,
  main = paste("Histogram of ", xlab), ..., colorQuantile = c("GREY",
  "YELLOW", "ORANGE", "DARK GREEN", "ORANGE", "YELLOW", "GREY"),
  colorProbability = c(1, 0.95, 0.75, 0.55, 0.45, 0.25, 0.05),
  resultName = NULL)
}
\arguments{
\item{x}{An object of class \code{mcSimulation}.}

\item{...}{Further arguments #ToDo}
}
\description{
This function plots the histogram of the results of \code{mcSimulation}.
}
\seealso{
\code{\link{mcSimulation}}, \code{\link{hist}}
}

