% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{paramtnormci}
\alias{paramtnormci}
\title{Return parameters of truncated normal random distribution based on a confidence interval.}
\usage{
paramtnormci(p, ci, lowerTrunc = -Inf, upperTrunc = Inf,
  relativeTolerance = 0.05, method = "numeric")
}
\arguments{
\item{p}{numeric vector; probabilities of upper and lower bound of the corresponding
confidence interval.}

\item{ci}{\code{numeric} vector; lower and upper bound of the  confidence interval.}

\item{lowerTrunc}{\code{numeric}; lower truncation point of the distribution.}

\item{upperTrunc}{\code{numeric}; upper truncation point of the distribution.}

\item{relativeTolerance}{\code{numeric}; the relative tolerance level of deviation of the generated confidence
interval from the specified interval.}

\item{method}{The method to calculate the parameters. Default is \code{"numeric"}.}
}
\description{
This function calculates the distribution parameters of a truncated normal distribution
from an arbitrary confidence interval.
}
\details{
#ToDo
}

