% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paramtruncnormci.R
\name{paramtruncnormci}
\alias{paramtruncnormci}
\title{Return parameters of truncated normal distribution based on a confidence interval.}
\usage{
paramtruncnormci(p, ci, lowerTrunc = -Inf, upperTrunc = Inf,
  relativeTolerance = 0.05, method = "numeric")
}
\arguments{
\item{p}{\code{numeric} 2-dimensional vector; probabilities of upper and lower bound of the corresponding
confidence interval.}

\item{ci}{\code{numeric} 2-dimensional vector; lower and upper bound of the  confidence interval.}

\item{lowerTrunc}{\code{numeric}; lower truncation point of the distribution (>= \code{-Inf}).}

\item{upperTrunc}{\code{numeric}; upper truncation point of the distribution (<= \code{Inf}).}

\item{relativeTolerance}{\code{numeric}; the relative tolerance level of deviation of the generated confidence
interval from the specified interval. If this deviation is greater than \code{relativeTolerance} a warning is given.}

\item{method}{The method to calculate the parameters. Default is \code{"numeric"}.}
}
\value{
A list with elements \code{mean} and \code{sd}.
}
\description{
This function calculates the distribution parameters, i.e. \code{mean} and \code{sd}, of a truncated normal distribution
from an arbitrary confidence interval.
}
\details{
For details of the truncated normal distribution see \code{\link[truncnorm]{truncnorm}}.

 p-\% confidence interval: \eqn{[c_l, c_u]}
 \deqn{tN(\mu,\sigma)}
 \deqn{\int_{-\infty}^{c_l} d tN(\mu,\sigma) = p_l}{\ifelse{html}{&int;}{int}}
 \deqn{\int_{-\infty}^{c_u} d tN(\mu,\sigma) = p_u}
 \eqn{p_l - p_u = p} and \eqn{p_l = 1- p_u}
 \if{latex}{\eqn{\int_{-\infty}^{c_l} d tN(\mu,\sigma) = p_l}}\if{html}{\out{<MATH>&int;{{su|b=-&infin;|p=c{{su|b=l}};}};</MATH>}}

 \if{html}{\out{&int;</sup><sub>-&infin;</sub><sup>c<sub>l</sub>}}\cr
}
\section{Warning}{

  This method has not been tested systematically!
}
\seealso{
\code{\link[truncnorm]{truncnorm}}
}

