% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.mcSimulation.R
\name{print.mcSimulation}
\alias{print.mcSimulation}
\title{Print Basic Results from Monte Carlo Simulation.}
\usage{
\method{print}{mcSimulation}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mcSimulation}.}

\item{...}{Further arguments to be passed to \code{\link{print.data.frame}}.}
}
\description{
This function prints basic results from Monte Carlo simulation  and returns it invisible.
}
\seealso{
\code{\link{mcSimulation}}, \code{\link{print.data.frame}}
}
