% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{print.summary.decisionAnalysis}
\alias{print.summary.decisionAnalysis}
\title{Print the Summarized Decsion Analysis Results..}
\usage{
\method{print}{summary.decisionAnalysis}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.decisionAnalysis}.}

\item{...}{Further arguments #ToDo}
}
\description{
This function prints the summary of of \code{decisionAnalysis} obtained by \code{\link{summary.decisionAnalysis}}.
}
\seealso{
\code{\link{decisionAnalysis}}
}

