% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{print.summary.mcSimulation}
\alias{print.summary.mcSimulation}
\title{Print the Summary of a Monte Carlo Simulation.}
\usage{
\method{print}{summary.mcSimulation}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mcSimulation}.}

\item{...}{Further arguments #ToDo}
}
\description{
This function prints the summary of of \code{mcSimulation} obtained by \code{\link{summary.mcSimulation}}.
}
\seealso{
\code{\link{mcSimulation}}
}

