% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/welfareDecisionAnalysis.R
\name{print.summary.welfareDecisionAnalysis}
\alias{print.summary.welfareDecisionAnalysis}
\title{Print the Summarized Decsion Analysis Results..}
\usage{
\method{print}{summary.welfareDecisionAnalysis}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.welfareDecisionAnalysis}.}

\item{...}{Further arguments #ToDo}
}
\description{
This function prints the summary of of \code{welfareDecisionAnalysis} obtained by \code{\link{summary.welfareDecisionAnalysis}}.
}
\seealso{
\code{\link{welfareDecisionAnalysis}}
}

