% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/r0_1norm90ci_numeric.R
\name{r0_1norm90ci_numeric}
\alias{r0_1norm90ci_numeric}
\title{Generate  normal random numbers truncated to \eqn{[0,1]} based on the 90\%-confidence interval.}
\usage{
r0_1norm90ci_numeric(n, lower, upper, relativeTolerance = 0.05)
}
\arguments{
\item{n}{Number of generated observations.}

\item{lower}{\code{numeric}; lower bound of the 90\% confidence intervall.}

\item{upper}{\code{numeric}; upper bound of the 90\% confidence intervall.}

\item{relativeTolerance}{\code{numeric}; the relative tolerance level of deviation of the generated confidence
interval from the specified interval.}
}
\description{
This function generatesnormal random numbers truncated to \eqn{[0,1]} based on the 90\% confidence interval
calculating the distribution parameter numerically from the  90\%-confidence interval.
}
\details{
#ToDo
}

