% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/random.R
\name{random.default}
\alias{random.default}
\title{Generate random numbers based on the first two moments of a certain probability distribution.}
\usage{
\method{random}{default}(rho = list(distribution_type, mean, sd), n, method,
  ...)
}
\arguments{
\item{rho}{list; Distribution to be randomly sampled.}

\item{n}{Number of generated observations}

\item{method}{Particular method to be used for random number generation.}

\item{...}{Optional arguments to be passed to the particular random number
generating function.}
}
\description{
This function generates random numbers for general multivariate
distributions that can be characterized by the joint first two moments, viz.
the mean and covariance.
}

