% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/random.R
\name{random}
\alias{random}
\title{Generate random numbers for a certain probability distribution.}
\usage{
random(rho, n, method, ...)
}
\arguments{
\item{rho}{Distribution to be randomly sampled.}

\item{n}{Number of generated observations.}

\item{method}{Particular method to be used for random number generation.}

\item{...}{Optional arguments to be passed to the particular random number
generating function.}
}
\description{
This function generates multivariate random numbers for general multivariate
distributions.
}

