% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{random_estimate_1d}
\alias{random_estimate_1d}
\title{Generate univariate random numbers based on an estimate.}
\usage{
random_estimate_1d(rho, n, method = "calculate", ...)
}
\arguments{
\item{rho}{\code{estimate} object; Univariate distribution to be randomly sampled.}

\item{n}{Number of generated observations}

\item{method}{Particular method to be used for random number generation.}

\item{...}{Optional arguments to be passed to the particular random number
generating function.}
}
\description{
This function generates random numbers for general univariate
distributions.
}
\details{
\code{method} can be either \code{"calculate"} (the default) or \code{"fit"}.
}

