% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/random_state.R
\name{random_state}
\alias{random_state}
\title{Draw a random state for a categorical variable}
\usage{
random_state(states, probs)
}
\arguments{
\item{states}{character vector containing state names.}

\item{probs}{numeric vector containing probabilities for the states. If these do
not add up to 1, they are automatically normalized.}
}
\value{
one of the states, drawn randomly according to the specified probabilities.
}
\description{
This function draws a sample from a user-defined frequency distribution for a categorical
variable.
}
\examples{

random_state(states=c("very low","low","medium","high","very high"),
  probs=c(1,1,2,1,1))

}
\author{
Eike Luedeling
}
\keyword{"conditional}
\keyword{"random}
\keyword{CPTs}
\keyword{probability}
\keyword{state"}
\keyword{table"}
