% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdistq_fit.R
\name{rdistq_fit}
\alias{rdistq_fit}
\title{Generate univariate random numbers based on quantiles.}
\usage{
rdistq_fit(distribution, n, percentiles = c(0.05, 0.5, 0.95), quantiles)
}
\arguments{
\item{distribution}{A character string that defines the univariate distribution
to be randomly sampled.}

\item{n}{Number of generated observations.}

\item{percentiles}{Numeric vector giving the percentiles.}

\item{quantiles}{Numeric vector giving the quantiles.}
}
\value{
ToDo
}
\description{
This function generates random numbers for a set of univariate distributions based on the
distribution quantiles. Internally, this is achieved by fitting the distribution function
to the given quantiles using \code{\link[rriskDistributions]{rriskFitdist.perc}}.
}
\details{
The follwing table shows the available distributions and their identification as a character string:
 \tabular{lll}{
 \bold{Identification} \tab  \bold{Distribution} \tab \bold{Number of quantiles}\cr
 \code{\link[=Normal]{norm}}       \tab Normal distribution  \tab >=2  \cr
 \code{\link[=Beta]{beta}}         \tab Beta distribution    \tab ToDo \cr
 \code{cauchy}     \tab ToDo \tab ToDo\cr
 \code{logis}      \tab ToDo \tab ToDo\cr
 \code{t}          \tab ToDo \tab ToDo\cr
 \code{chisq}      \tab ToDo \tab ToDo\cr
 \code{chisqnc}    \tab ToDo: implement? \tab ToDo\cr
 \code{exp}        \tab ToDo \tab ToDo\cr
 \code{f}          \tab ToDo \tab ToDo\cr
 \code{gamma}      \tab ToDo \tab ToDo\cr
 \code{lnorm}      \tab ToDo  \tab ToDo\cr
 \code{unif}       \tab ToDo \tab ToDo\cr
 \code{weibull}    \tab ToDo \tab ToDo\cr
 \code{\link[mc2d:triangular]{triang}}   \tab Triangular Distribution, Note: package \code{mc2d} needed. \tab ToDo\cr
 \code{\link[eha:Gompertz]{gompertz}}   \tab Gompertz distribution, Note: package \code{eha} needed.\tab ToDo\cr
 \code{\link[mc2d]{pert}}       \tab The (modified) PERT distribution,  Note: package \code{mc2d} needed. \tab ToDo\cr
 \code{\link[msm]{tnorm}}      \tab Truncated normal distribution \tab ToDo
 }
The default for \code{percentiles} is 0.05, 0.5 and 0.95, so for the default,
the quantiles argument should be a vector with 3 elements. If this is to be longer,
the percentiles argument has to be adjusted to match the length of quantiles.
}

