% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{rmvnorm90ci_exact}
\alias{rmvnorm90ci_exact}
\title{Generate normal distributed multivariate random numbers based on the 90\%-confidence interval.}
\usage{
rmvnorm90ci_exact(n, lower, upper, correlationMatrix)
}
\arguments{
\item{n}{Number of generated observations.}

\item{lower}{\code{numeric} vector; lower bound of the 90\% confidence intervall.}

\item{upper}{\code{numeric} vector; upper bound of the 90\% confidence intervall.}

\item{correlationMatrix}{\code{numeric} symmetric matrix; correlation matrix;
In particular, all diagonal elements must be equal to 1.}
}
\description{
This function generates normal distributed multivariate random numbers
based on the 90\%-confidence interval.
}

