% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rposnorm90ci.R
\name{rposnorm90ci}
\alias{rposnorm90ci}
\title{Generate positive normal random numbers based on the 90\%-confidence interval.}
\usage{
rposnorm90ci(n, lower, median = mean(c(lower, upper)), upper,
  method = "numeric", relativeTolerance = 0.05, ...)
}
\arguments{
\item{n}{Number of generated observations.}

\item{lower}{\code{numeric}; lower bound of the 90\% confidence intervall.}

\item{median}{if \code{NULL}: truncated normal is fitted only to lower and upper value of the
confidence interval; if \code{numeric}: truncated normal is fitted on the confidence interval
and the median simultaneously. For details cf. below. This option is only relevant if
\code{method="fit"}.}

\item{upper}{\code{numeric}; upper bound of the 90\% confidence intervall.}

\item{method}{method used to determine the parameters of the positive normal; possible methods
are \code{"numeric"} (the default) and \code{"fit"}.}

\item{relativeTolerance}{\code{numeric}; the relative tolerance level of deviation of the
generated confidence interval from the specified interval. If this deviation is greater than
\code{relativeTolerance} a warning is given.}

\item{...}{further parameters to be passed to \code{\link{paramtnormci_numeric}} or
\code{\link{paramtnormci_fit}}, respectively.}
}
\description{
This function generates positive normal random numbers based on the 90\% confidence interval
calculating the distribution parameter numerically from the  90\%-confidence interval or via a
fit on the 90\%-confidence interval. The fit might include the median or not.
}
\details{
For the implementation of \code{method="numeric"}: \code{\link{paramtnormci_numeric}};
 for the implementation of \code{method="fit"}: \code{\link{paramtnormci_fit}}.
}
\seealso{
For the implementation of \code{method="numeric"}: \code{\link{paramtnormci_numeric}};
 for the implementation of \code{method="fit"}: \code{\link{paramtnormci_fit}}.
}

