% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{rposnorm90ci_iter}
\alias{rposnorm90ci_iter}
\title{Generate positive normal random numbers based on the 90\%-confidence interval.}
\usage{
rposnorm90ci_iter(n, lower, upper, relativeTolerance = 0.05, maxIter = 40)
}
\arguments{
\item{n}{Number of generated observations.}

\item{lower}{\code{numeric}; lower bound of the 90\% confidence intervall.}

\item{upper}{\code{numeric}; upper bound of the 90\% confidence intervall.}

\item{relativeTolerance}{\code{numeric}; the relative tolerance level of deviation of the generated confidence
interval from the specified interval.}

\item{maxIter}{\code{numeric}; maximum number of iterations.}
}
\description{
This function generates positive normal random numbers based on the 90\% confidence interval
using an iteration algorithm.
}
\details{
The generation of random numbers is repeated until the generated 90\% - confidence interval is
close enough to the desired value.
}

