% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{sort.summary.eviSimulation}
\alias{sort.summary.eviSimulation}
\title{Sort Summarized EVI Simulation Results..}
\usage{
\method{sort}{summary.eviSimulation}(x, decreasing = TRUE, ...,
  along = row.names(x$summary$evi)[[1]])
}
\arguments{
\item{x}{An object of class \code{summary.eviSimulation}.}

\item{decreasing}{logical; if the evi should be sorted in decreasing order.}

\item{...}{Further arguments #ToDo}

\item{along}{character; the name of the valuation variable along which evi
should be sorted.}
}
\value{
An object of class \code{summary.eviSimulation}.
}
\description{
Sort summarized EVI simulation results according to their EVI.
}
\seealso{
\code{\link{eviSimulation}}, \code{\link{summary.eviSimulation}}, \code{\link[base]{sort}}
}

