% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{summary.decisionAnalysis}
\alias{summary.decisionAnalysis}
\title{Summarize Decsion Analysis Results.}
\usage{
\method{summary}{decisionAnalysis}(object, ..., digits = max(3,
  getOption("digits") - 3))
}
\arguments{
\item{object}{An object of class \code{decisionAnalysis}.}

\item{...}{Further arguments #ToDo}
}
\value{
An object of class \code{summary.decisionAnalysis}.
}
\description{
summary.decisionAnalysis produces result summaries of the results of decision analysis
 simulation obtained by the function \code{\link{decisionAnalysis}}.
}
\seealso{
\code{\link{decisionAnalysis}}, \code{\link{print.summary.decisionAnalysis}}
}

