% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/welfareDecisonAnalysis.R
\name{summary.welfareDecisionAnalysis}
\alias{summary.welfareDecisionAnalysis}
\title{Summarize Decsion Analysis Results.}
\usage{
\method{summary}{welfareDecisionAnalysis}(object, ..., digits = max(3,
  getOption("digits") - 3))
}
\arguments{
\item{object}{An object of class \code{welfareDecisionAnalysis}.}

\item{...}{Further arguments #ToDo}
}
\value{
An object of class \code{summary.welfareDecisionAnalysis}.
}
\description{
summary.welfareDecisionAnalysis produces result summaries of the results of decision analysis
 simulation obtained by the function \code{\link{welfareDecisionAnalysis}}.
}
\seealso{
\code{\link{welfareDecisionAnalysis}}, \code{\link{print.summary.welfareDecisionAnalysis}}
}

