% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as.data.frame.mcSimulation.R
\name{as.data.frame.mcSimulation}
\alias{as.data.frame.mcSimulation}
\title{Coerce Monte Carlo simulation results to a data frame.}
\usage{
\method{as.data.frame}{mcSimulation}(
  x,
  row.names = NULL,
  optional = FALSE,
  ...,
  stringsAsFactors = NA
)
}
\arguments{
\item{x}{An object of class \code{mcSimulation}.}

\item{row.names}{\code{NULL} or a character vector giving the row
    names for the data frame.  Missing values are not allowed.}

\item{optional}{logical. If \code{TRUE}, setting row names and
    converting column names (to syntactic names: see
    \code{\link[base]{make.names}}) is optional.  Note that all of \R's
    \pkg{base} package \code{as.data.frame()} methods use
    \code{optional} only for column names treatment, basically with the
    meaning of \code{\link[base]{data.frame}(*, check.names = !optional)}.
    See also the \code{make.names} argument of the \code{matrix} method.}

\item{...}{additional arguments to be passed to or from methods.}

\item{stringsAsFactors}{logical: should the character vector be converted
    to a factor?}
}
\description{
Coerces Monte Carlo simulation results to a data frame.
}
\seealso{
\code{\link{as.data.frame}}
}
