% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{estimate}
\alias{estimate}
\title{Create an Estimate Object}
\usage{
estimate(..., correlation_matrix = NULL)
}
\arguments{
\item{...}{arguments that can be coerced to a data frame comprising the base of the estimate.}

\item{correlation_matrix}{numeric matrix containing the correlations of the variables.}
}
\value{
An object of type \code{estimate} which is a list whith components \code{base} and \code{correlation_matrix}.
\code{base} is a \code{\link{data.frame}} with mandatory column \code{distribution}. The \code{\link{row.names}} are the
names of the variables. \code{correlation_matrix} is a symmetric matrix with row and column names being the subset of
the variables supplied in \code{base} which are correlated. Its elements are the corresponding correlations.
}
\description{
This function creates an object of class \code{estimate}. #ToDo: detailed description
#ToDo: Implement characterization of distribution by mean and sd. Eventually, also by other quantiles.
}
\details{
The parameters in \code{...} provide the base information of an estimate.
\subsection{The structure of the estimate base information (mandatory)}{
   Mandatory columns:
   \tabular{lll}{
     Column name         \tab  R-type    \tab Explanation\cr
     \code{distribution} \tab  \code{character} \tab  Distribution types \cr
     \code{variable}     \tab  \code{character} \tab  Variable names
   }
}
}
\seealso{
\code{\link{row.names.estimate}}, \code{\link{names.estimate}}, \code{\link{corMat}}, \code{\link{estimate_read_csv}},
\code{\link{estimate_write_csv}}, \code{\link{random.estimate}}
}

