% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate.R
\name{names.estimate}
\alias{names.estimate}
\title{Return the column names of an \code{estimate} object.}
\usage{
\method{names}{estimate}(x)
}
\arguments{
\item{x}{an \code{\link{estimate}} object.}
}
\description{
This function returns the column names of an \code{\link{estimate}} object which is identical to
\code{names(x$base)}.
}
\seealso{
\code{\link{estimate}}, \code{\link{row.names.estimate}}, \code{\link{corMat.estimate}}
}

