% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/paramposnorm90ci.R
\name{paramposnorm90ci}
\alias{paramposnorm90ci}
\title{Return parameters of positive normal random distribution based on the 90\%-confidence interval.}
\usage{
paramposnorm90ci(lower, upper, relativeTolerance = 0.05, method = "numeric")
}
\arguments{
\item{lower}{\code{numeric}; lower bound of the 90\% confidence intervall.}

\item{upper}{\code{numeric}; upper bound of the 90\% confidence intervall.}

\item{relativeTolerance}{\code{numeric}; the relative tolerance level of deviation of the generated confidence
interval from the specified interval.}

\item{method}{The method to calculate the parameters. Default is \code{"numeric"}.}
}
\description{
This function calculates the distribution parameters from the  90\%-confidence interval.
}
\details{
#ToDo
}

