% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_cashflow.R
\name{plot_cashflow}
\alias{plot_cashflow}
\title{Cashflow plot for Monte Carlo simulation results}
\usage{
plot_cashflow(
  mcSimulationResults,
  cashflow_var_name,
  xlabel = "Timeline of intervention",
  ylabel = "Cashflow",
  color_25_75 = "grey40",
  color_5_95 = "grey70",
  color_median = "blue"
)
}
\arguments{
\item{mcSimulationResults}{is a data frame of Monte Carlo simulations of cashflow outputs (in wide format, see example). Usually the "MCSimulationResults.csv" file generated by \code{\link[decisionSupport]{decisionSupport}} function.}

\item{cashflow_var_name}{is a character string representing the variable name used to define cashflow in the returned list of outputs from the \code{\link[decisionSupport]{decisionSupport}} function.}

\item{xlabel}{is a character string representing the title of the timeline of the intervention to be printed on the x axis in quotes.}

\item{ylabel}{is a character string representing the title of the units of the cashflow to be printed on the y axis.}

\item{color_25_75}{is a character string referring to the color name for the shade fill of the 25-75\% quantile from the grDevices colors. The default is "grey40".}

\item{color_5_95}{is a character string referring to the shade fill of the 5-95\% quantile from the grDevices colors. The default is "grey70".}

\item{color_median}{is a character string  referring to the color name for the median line from the grDevices colors. The default is  "blue".}
}
\description{
Creates a cashflow plot of the returned list of related outputs from the \code{\link[decisionSupport]{decisionSupport}} function using \code{\link{ggplot2}}
}
\details{
This function automatically defines quantiles (5 to 95\% and 25 to 75\%) as well as a value for the median.
}
\examples{

#Generate random dataset of 200 Monte Carlo simulations for cashflow

test_data <- data.frame(replicate(20, sample(-10000:10000, 200, rep = TRUE)))

#Make a name for the cashflow variables with time append (as is done in the decisionSupport function)

names(test_data) <-  gsub(x = names(test_data),
                          pattern = "X",
                          replacement = "cashflow")


plot_cashflow(mcSimulationResults = test_data, cashflow_var_name = "cashflow",
              xlabel = "Years with intervention",
              ylabel = "Annual cashflow in USD", 
              color_25_75 = "green4", color_5_95 = "green1",
              color_median = "red")
  
}
\references{
Lanzanova Denis, Cory Whitney, Keith Shepherd, and Eike Luedeling. “Improving Development Efficiency through Decision Analysis: Reservoir Protection in Burkina Faso.” Environmental Modelling & Software 115 (May 1, 2019): 164–75. \url{https://doi.org/10.1016/j.envsoft.2019.01.016}.
}
\keyword{Monte-Carlo}
\keyword{cashflow}
\keyword{decision-analysis}
\keyword{decisionSupport}
\keyword{risk}
\keyword{uncertainty}
