% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_evpi.R
\name{plot_evpi}
\alias{plot_evpi}
\title{Visualizing the results of Expected Value of Perfect Information (EVPI) analysis for various types of Monte Carlo simulation results}
\usage{
plot_evpi(
  EVPIresults,
  input_table = NULL,
  decision_vars,
  new_names = NULL,
  unit = NULL,
  bar_color = "cadetblue",
  ...
)
}
\arguments{
\item{EVPIresults}{are the results of the \code{\link[decisionSupport:multi_EVPI]{multi_EVPI}} function}

\item{input_table}{is a data frame with at least two columns named 'variable' and 'label'. The 'variable column should have one entry for the name of each variable contained in any of the plots. In preparing the figure, the function will replace the variable names with the labels. If the label is missing then the plot will show 'NA' in the place of the variable name. Default is NULL and uses the original variable names.}

\item{decision_vars}{are the names of the decision variables in the output of the \code{\link[decisionSupport:mcSimulation]{mcSimulation}} function}

\item{new_names}{are the reformatted replacement names of the decision variables in the output of the \code{\link[decisionSupport:mcSimulation]{mcSimulation}} function}

\item{unit}{is the symbol to display before the evpi value on the x axis. It accepts text or (many) unicode formatted symbol text}

\item{...}{accepts arguments to be passed to \code{\link[ggplot2:theme]{ggplot::theme}}}
}
\description{
Plotting the Expected Value of Perfect Information (EVPI) of Monte Carlo outputs
}
\examples{

# Create a data.frame

montecarlo <- data.frame(indep1 = rnorm(1000, sd = 50, mean = 100), indep2 = rnorm(1000, sd = 100, mean = 100))
montecarlo[, 'output1'] <- montecarlo[, 'indep1'] * montecarlo[, 'indep2']
montecarlo[, 'output2'] <- (montecarlo[, 'indep1'] * (montecarlo[, 'indep2']) + 10)

# Run the multi_EVPI function 
 
results_all <- multi_EVPI(montecarlo,"output1")

plot_evpi(results_all, decision_vars = c("output1", "output2"), 
new_names = c("Decision option 1", "Decision option 2"))

}
\references{
Do, Hoa, Eike Luedeling, and Cory Whitney. “Decision Analysis of Agroforestry Options Reveals Adoption Risks for Resource-Poor Farmers.” Agronomy for Sustainable Development 40, no. 3 (June 2020): 20. \url{https://doi.org/10.1007/s13593-020-00624-5}
Lanzanova, Denis, Cory Whitney, Keith Shepherd, and Eike Luedeling. “Improving Development Efficiency through Decision Analysis: Reservoir Protection in Burkina Faso.” Environmental Modelling & Software 115 (May 1, 2019): 164–75. \url{https://doi.org/10.1016/j.envsoft.2019.01.016}
Luedeling, Eike, and Keith Shepherd. “Decision-Focused Agricultural Research.” Solutions 7, no. 5 (2016): 46–54. \url{https://www.thesolutionsjournal.com/article/decision-focused-agricultural-research/}
}
\keyword{EVPI}
\keyword{Monte-Carlo}
\keyword{decision-analysis}
\keyword{decisionSupport}
\keyword{net-present-value}
\keyword{perfect-information}
\keyword{risk}
\keyword{uncertainty}
