% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{print.mcSimulation}
\alias{print.mcSimulation}
\title{Print Basic Results from Monte Carlo Simulation.}
\usage{
\method{print}{mcSimulation}(x, ...)
}
\arguments{
\item{x}{An object of class \code{mcSimulation}.}

\item{...}{Further arguments #ToDo}
}
\description{
This function prints basic results from Monte Carlo simulation  and returns it invisible.
}
\seealso{
\code{\link{mcSimulation}}
}

