% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviSimulation.R
\name{print.summary.eviSimulation}
\alias{print.summary.eviSimulation}
\title{Print the Summarized EVI Simulation Results.}
\usage{
\method{print}{summary.eviSimulation}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.eviSimulation}.}

\item{...}{Further arguments to be passed to \code{\link{print.default}} and \ifelse{latex}{\cr}{ }
\code{\link{print.summary.welfareDecisionAnalysis}}.}
}
\description{
This function prints the summary of \code{eviSimulation} generated by 
\code{\link{summary.eviSimulation}}.
}
\seealso{
\code{\link{eviSimulation}}, \code{\link{print.summary.welfareDecisionAnalysis}}.
}
