% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{print.summary.eviSimulation}
\alias{print.summary.eviSimulation}
\title{Print the Summarized EVI Simulation Results.}
\usage{
\method{print}{summary.eviSimulation}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.eviSimulation}.}

\item{...}{Further arguments #ToDo}
}
\description{
This function prints the summary of of \code{eviSimulation} obtained by \code{\link{summary.eviSimulation}}.
}
\seealso{
\code{\link{eviSimulation}}
}

