% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/welfareDecisionAnalysis.R
\name{print.summary.welfareDecisionAnalysis}
\alias{print.summary.welfareDecisionAnalysis}
\title{Print the summarized Welfare Decision Analysis results.}
\usage{
\method{print}{summary.welfareDecisionAnalysis}(x, ...)
}
\arguments{
\item{x}{An object of class \code{summary.welfareDecisionAnalysis}.}

\item{...}{Further arguments to \code{\link{print.data.frame}}.}
}
\description{
This function prints the summary of a Welfare Decision Analysis generated by \ifelse{latex}{\cr}{ }
\code{\link{summary.welfareDecisionAnalysis}}.
}
\seealso{
\code{\link{welfareDecisionAnalysis}}, \code{\link{summary.welfareDecisionAnalysis}},
  \code{\link{print.data.frame}}.
}

