% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/estimate.R
\name{random.estimate}
\alias{random.estimate}
\title{Generate Random Numbers for an Estimate.}
\usage{
\method{random}{estimate}(rho, n, method = "calculate", ...)
}
\arguments{
\item{rho}{\code{estimate} object; Multivariate distribution to be randomly sampled.}

\item{n}{Number of generated observations}

\item{method}{Particular method to be used for random number generation.}

\item{...}{Optional arguments to be passed to the particular random number
generating function.}
}
\description{
This function generates random numbers for general multivariate
distributions that are defined as an \code{\link{estimate}}.
}
\details{
\subsection{Generation of uncorrelated components}{
		Implementation: \code{\link{random_estimate_1d}}

	}
	\subsection{Generation of correlated components}{
		Implementation: \code{\link{rmvnorm90ci_exact}}
	}
}
\examples{
variable=c("revenue","costs")
	distribution=c("norm","norm")
 lower=c(10000,  5000)
 upper=c(100000, 50000)
 estimateObject<-estimate(variable, distribution, lower, upper)
 x<-random(rho=estimateObject, n=10000)
 apply(X=x, MARGIN=2, FUN=quantile, probs=c(0.05, 0.95))
 cor(x)
 colnames(x)
 summary(x)
 hist(x[,"revenue"])
 hist(x[,"costs"])
}
\seealso{
\code{\link{estimate}}
}

