% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/rdist90ci_exact.R
\name{rdist90ci_exact}
\alias{rdist90ci_exact}
\title{Generate univariate random numbers based on the 90\%-confidence interval.}
\usage{
rdist90ci_exact(distribution, n, lower, upper)
}
\arguments{
\item{distribution}{\code{character}; A character string that defines the univariate distribution
to be randomly sampled.}

\item{n}{Number of generated observations.}

\item{lower}{\code{numeric}; lower bound of the 90\% confidence intervall.}

\item{upper}{\code{numeric}; upper bound of the 90\% confidence intervall.}
}
\description{
This function generates random numbers for general univariate
distributions based on the 90\% confidence interval.
}
\details{
The follwing table shows the available distributions and their identification as a character string:
 \tabular{ll}{
 Distribution encoding \tab  Distribution\cr
 \code{const}       \tab ToDo \cr
 \code{norm}         \tab ToDo \cr
 \code{pos_norm}     \tab ToDo \cr
 \code{norm_0_1}     \tab ToDo \cr
 \code{pois}        \tab ToDo \cr
 \code{binom}       \tab ToDo \cr
 \code{unif}        \tab ToDo \cr
 \code{lnorm}        \tab ToDo \cr
 \code{lnorm_lim2}   \tab ToDo
 }
}

