% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rmvnorm90ci_exact.R
\name{rmvnorm90ci_exact}
\alias{rmvnorm90ci_exact}
\title{90\%-confidence interval multivariate normal random number generation.}
\usage{
rmvnorm90ci_exact(n, lower, upper, correlationMatrix)
}
\arguments{
\item{n}{\code{integer}: Number of observations to be generated.}

\item{lower}{\code{numeric vector}: lower bound of the 90\% confidence interval.}

\item{upper}{\code{numeric vector}: upper bound of the 90\% confidence interval.}

\item{correlationMatrix}{\code{numeric matrix}: symmetric matrix which is the correlation matrix of the 
multivariate normal distribution. In particular, all diagonal elements must be equal to 1.}
}
\description{
This function generates normally distributed multivariate random numbers which parameters are 
determined by the 90\%-confidence interval. The calculation of \code{mean} and \code{sd} is 
exact.
}
\seealso{
\code{\link{random}}, \code{\link[mvtnorm]{rmvnorm}}
}

