% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{row.names.estimate}
\alias{row.names.estimate}
\title{Return the variable names of an \code{estimate} object.}
\usage{
\method{row.names}{estimate}(x)
}
\arguments{
\item{x}{an \code{\link{estimate}} object.}
}
\description{
This function returns the variable names of an \code{\link{estimate}} object which is identical to
\code{row.names(x$base)}.
}
\seealso{
\code{\link{estimate}}, \code{\link{names.estimate}}, \code{\link{corMat.estimate}}
}

