% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{rposnorm90ci_numeric}
\alias{rposnorm90ci_numeric}
\title{Generate positive normal random numbers based on the 90\%-confidence interval.}
\usage{
rposnorm90ci_numeric(n, lower, upper, relativeTolerance = 0.05)
}
\arguments{
\item{n}{Number of generated observations.}

\item{lower}{\code{numeric}; lower bound of the 90\% confidence intervall.}

\item{upper}{\code{numeric}; upper bound of the 90\% confidence intervall.}

\item{relativeTolerance}{\code{numeric}; the relative tolerance level of deviation of the generated confidence
interval from the specified interval.}
}
\description{
This function generates positive normal random numbers based on the 90\% confidence interval
calculating the distribution parameter numerically from the  90\%-confidence interval.
}
\details{
#ToDo
}

