% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eviSimulation.R
\name{sort.summary.eviSimulation}
\alias{sort.summary.eviSimulation}
\title{Sort Summarized EVI Simulation Results..}
\usage{
\method{sort}{summary.eviSimulation}(x, decreasing = TRUE, ...,
  along = row.names(x$summary$evi)[[1]])
}
\arguments{
\item{x}{An object of class \code{summary.eviSimulation}.}

\item{decreasing}{\code{logical}: if the EVI should be sorted in decreasing order.}

\item{...}{currently not used}

\item{along}{\code{character}: the name of the valuation variable along which the EVI 
should be sorted.}
}
\value{
An object of class \code{summary.eviSimulation}.
}
\description{
Sort summarized EVI simulation results according to their EVI.
}
\seealso{
\code{\link{eviSimulation}}, \code{\link{summary.eviSimulation}}, \code{\link[base]{sort}}
}
