% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/eviSimulation.R
\name{summary.eviSimulation}
\alias{summary.eviSimulation}
\title{Summarize EVI Simulation Results..}
\usage{
\method{summary}{eviSimulation}(object, ..., digits = max(3,
  getOption("digits") - 3))
}
\arguments{
\item{object}{An object of class \code{eviSimulation}.}

\item{...}{Further arguments #ToDo}
}
\value{
An object of class \code{summary.eviSimulation}.
}
\description{
summary.eviSimulation produces result summaries of the results of Expected Value of
 Information (EVI) simulation obtained by the function \code{\link{eviSimulation}}.
}
\seealso{
\code{\link{eviSimulation}}, \code{\link{print.summary.eviSimulation}}
}

