% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mcSimulation.R
\name{summary.mcSimulation}
\alias{summary.mcSimulation}
\title{Summarize Results from Monte Carlo Simulation.}
\usage{
\method{summary}{mcSimulation}(object, ..., digits = max(3,
  getOption("digits") - 3), variables.y = names(object$y), variables.x = if
  (classicView) names(object$x), classicView = FALSE, probs = c(0, 0.1,
  0.25, 0.5, 0.75, 0.9, 1))
}
\arguments{
\item{object}{An object of class \code{mcSimulation}.}

\item{...}{Further arguments passed to \code{\link{summary.data.frame}} (\code{classicView=TRUE}) or \code{\link{format}} (\code{classicView=FALSE}).}
}
\value{
An object of class \code{summary.mcSimulation}.
}
\description{
summary.mcSimulation produces result summaries of the results of a Monte Carlo simulation obtained by the function \code{\link{mcSimulation}}.
}
\seealso{
\code{\link{mcSimulation}}, \code{\link{print.summary.mcSimulation}}, \code{\link{summary.data.frame}}
}

