% Generated by roxygen2 (4.0.2.9000): do not edit by hand
\name{uncertaintyAnalysis}
\alias{uncertaintyAnalysis}
\title{Uncertainty Analysis Wrapper Function.}
\usage{
uncertaintyAnalysis(inputFileName, outputDirectory, modelFunction,
  NumberofSimulations, randomMethod = "calculate",
  functionSyntax = "globalNames", write_table = TRUE, indicators = FALSE,
  log_scales = FALSE, oldInputStandard = FALSE)
}
\arguments{
\item{inputFileName}{Path to input csv file, which gives the input \code{\link{estimate}}.}

\item{outputDirectory}{Path were the result plots and tables are safed.}

\item{modelFunction}{The model function.}

\item{NumberofSimulations}{The number of Monte Carlo simulations to be performed.}

\item{randomMethod}{ToDo}

\item{functionSyntax}{ToDo}

\item{write_table}{\code{logical}; If the full Monte Carlo simulation results and PLSR results should be
written to file.}

\item{indicators}{\code{logical}; If indicator variables should be respected specially.}

\item{log_scales}{\code{logical}; If the scales in the pls plots should be logarithmic.}

\item{oldInputStandard}{\code{logical}; If the old input standard should be used
(\code{\link{estimate_read_csv_old}}).}
}
\description{
This function performs a Monte Carlo simulation from input files and analyses the results
via Partial Least Squares Regression (PLSR) and calculates the Variable Importance on Projection
(VIP). Results are safed as plots.
}
\seealso{
\code{\link{mcSimulation}}, \code{\link{estimate}}, \code{\link{estimate_read_csv}}
}

